+++ js/api.js
@@ -7,6 +7,8 @@ const setInitialState = () => {
   setInitialMapState();
 };
 
+// Функция получения данных с сервера
+
 const getData = (onSuccess) => {
   fetch('https://23.javascript.pages.academy/keksobooking/data')
     .then((response) => response.json())
@@ -18,13 +20,37 @@ const getData = (onSuccess) => {
     });
 };
 
+// Функция отправки данных на сервер
 
-// Обрабочик cброса данных формы и карты
+// const sendData = (onSuccess, onFail, body) => {
+//   fetch(
+//     'https://23.javascript.pages.academy/keksobooking',
+//     {
+//       method: 'POST',
+//       body,
+//     },
+//   )
+//     .then((response) => {
+//       if (response.ok) {
+//         onSuccess();
+//       }
+//       else {
+//         onFail();
+//       }
+//     })
+//     .catch(() => {
+//       onFail();
+//     });
+// };
 
-const resetButton = document.querySelector('.ad-form__reset');
-resetButton.addEventListener('click', (evt) => {
-  evt.preventDefault();
-  setInitialState();
-});
+// // Обрабочик cброса данных формы и карты
+//
+// const resetButton = document.querySelector('.ad-form__reset');
+// resetButton.addEventListener('click', (evt) => {
+//   evt.preventDefault();
+//   setInitialState();
+// });
 
-export {getData, setInitialState};
+export {getData};
+// export {sendData};
+export {setInitialState};
